Move the Media components out of the Embed namespace
The Embed namespace is not needed.
