+++ app/components/companion_windows_component.html.erb
@@ -74,7 +74,7 @@
       Share
     </header>
 
-    <%= render Embed::Media::EmbedThisFormComponent.new(viewer:) %>
+    <%= render Media::EmbedThisFormComponent.new(viewer:) %>
 
     <button data-action="click->companion-window#closeModal">Close</button>
   </dialog>
