+++ app/components/media/metadata_component.html.erb
