+++ app/components/media/metadata_component.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Media
+  class MetadataComponent < ViewComponent::Base
+    def initialize(viewer:)
+      @viewer = viewer
+    end
+
+    attr_reader :viewer
+  end
+end
