+++ app/components/media_component.html.erb
@@ -91,7 +91,7 @@
         </picture>
       </div>
 
-      <%= render Embed::Media::TagComponent.with_collection(resources_with_primary_file, druid:) %>
+      <%= render Media::TagComponent.with_collection(resources_with_primary_file, druid:) %>
   <% end %>
 
   <% component.with_dialog do %>
