+++ spec/components/media/embed_this_form_component_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe Embed::Media::EmbedThisFormComponent, type: :component do
+RSpec.describe Media::EmbedThisFormComponent, type: :component do
   let(:request) do
     Embed::Request.new(url: 'http://purl.stanford.edu/abc123')
   end
